3-connect-to-snowflake/11-data-viewer-to-cloud/requirements.txt
pandas
streamlit
plotly
pyvis
scipy
networkx
snowflake-snowpark-python